--- tests/e2e.test.ts
@@ -1,66 +1,0 @@
-import { describe, expect, test } from "bun:test";
-import { fileURLToPath } from "url";
-import { dirname, join } from "path";
-
-const entry = join(dirname(fileURLToPath(import.meta.url)), "..", "src", "index.ts");
-
-async function runCli(args: string[]) {
-  const proc = Bun.spawn(["bun", "run", entry, ...args], {
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-
-  const [stdout, stderr, exitCode] = await Promise.all([
-    new Response(proc.stdout).text(),
-    new Response(proc.stderr).text(),
-    proc.exited,
-  ]);
-
-  return {
-    stdout: stdout.trim(),
-    stderr: stderr.trim(),
-    exitCode,
-  };
-}
-
-describe("cli", () => {
-  test("hello prints the current text", async () => {
-    const result = await runCli(["hello"]);
-    expect(result.exitCode).toBe(0);
-    expect(result.stderr).toBe("");
-    expect(result.stdout).toBe("Hello, World!");
-  });
-
-  test("calc prints only the number result", async () => {
-    const result = await runCli(["calc", "2", "+", "3"]);
-    expect(result.exitCode).toBe(0);
-    expect(result.stderr).toBe("");
-    expect(result.stdout).toBe("5");
-  });
-
-  test("calc computes the modulo of two numbers", async () => {
-    const result = await runCli(["calc", "13", "%", "5"]);
-    expect(result.exitCode).toBe(0);
-    expect(result.stderr).toBe("");
-    expect(result.stdout).toBe("3");
-  });
-
-  test("calc fails with invalid operator", async () => {
-    const result = await runCli(["calc", "2", "^", "3"]);
-    expect(result.exitCode).not.toBe(0);
-    expect(result.stderr).toContain("Allowed choices are +, -, *, /, %.");
-  });
-
-  test("calc fails with invalid number", async () => {
-    const result = await runCli(["calc", "abc", "+", "3"]);
-    expect(result.exitCode).not.toBe(0);
-    expect(result.stderr).toContain("is not a number");
-  });
-
-  test("prints help information", async () => {
-    const result = await runCli(["--help"]);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toContain("hello");
-    expect(result.stdout).toContain("calc");
-  });
-});
